+++ R/reporter-zzz.R
@@ -14,9 +14,12 @@
 #'   If `FALSE`, the data frame will be copied and `get_testdata()` will return
 #'   the state of the data frame at the time `set_testdata()` was called.
 #' @return
-#'   * `set_testdata()` returns the previous test data frame.
+#'   * `set_testdata()` invisibly returns the previous test data. The test data
+#'   is returned as it was stored - if it was stored with `quosure = TRUE` it
+#'   will be returned as a quosure.
 #'   * `get_testdata()` returns the current test data frame.
-#'   * `with_testdata()` returns the input `data` for easy piping.
+#'   * `with_testdata()` and the test data pipe `%E>%` invisibly return the
+#'   input `data` for easy piping.
 #' @examples
 #' set_testdata(mtcars)
 #' head(get_testdata())
@@ -25,6 +28,10 @@
 #'   x <- get_testdata()
 #'   print(head(x))
 #' })
+#'
+#' mtcars %E>%
+#'   expect_base(mpg, TRUE) %E>%
+#'   expect_range(carb, 1, 8)
 #' @name global-data
 NULL
 
@@ -36,7 +43,7 @@ set_testdata <- function(data, quosure = TRUE) {
   old <- testdat_env$test_data
   if (quosure) data <- enquo(data)
   assign("test_data", data, testdat_env)
-  invisible(eval_tidy(old))
+  invisible(old)
 }
 
 #' @export
@@ -54,14 +61,20 @@ get_testdata <- function() {
 
 #' @export
 #' @rdname global-data
-#' @param code Code to execute.
+#' @param code Code to execute with the test data set to `data`.
 with_testdata <- function(data, code, quosure = TRUE) {
   old <- set_testdata(data, quosure = quosure)
-  on.exit(set_testdata(old), add = TRUE)
+  on.exit(set_testdata(old, quosure = FALSE), add = TRUE)
 
   force(code)
 
   invisible(data)
+}
+
+#' @rdname global-data
+#' @export
+`%E>%` <- function(data, code) {
+  with_testdata(data, code, quosure = FALSE)
 }
 
 data_reporter <- function() {